src package renamed. example package created. note test added.
